Separated Details and DetailCard, fixed responsiveness in Details
